R script for Pocket PlayLab challenge
The objective of this challenge is to analyse a data file.
Requirements & output
The datafile contains a cleaned sample of game-events from a given day.
The format of the file is a simple CSV file with the first line being a
header. The output should be a small analysis of the supplied events. It
should at least contain the following